pyproject.toml
[tool.poetry]
name = "study-buddy"
version = "0.1.0"
description = ""
authors = ["Casey <caseyjohnsonwv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.13"
langchain-community = "^0.0.29"
langchain-openai = "^0.1.1"
pypdf = "^4.1.0"
python-pptx = "^0.6.23"
gradio = "^4.23.0"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
